+++ src/index.ts
@@ -11,6 +11,7 @@ export { Routing, DependsOnMethod } from './routing';
 export { createServer, attachRouting } from './server';
 export { OpenAPI } from './open-api';
 export { OpenAPIError, DependsOnMethodError, RoutingError } from './errors';
+export { withMeta } from './metadata';
 
 import * as z from './extend-zod';
 import createHttpError from 'http-errors';